removed FirebaseAuth
