+++ lib/src/account_settings/view/menu.dart
@@ -1,34 +1,28 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_auth/firebase_auth.dart'; // This import is only to use FirebaseUser class
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import '../view_model/login_view_model.dart';
 
 class Menu extends StatelessWidget {
-  void _testDataBase() {
-    Firestore.instance
-        .collection('books')
-        .document()
-        .setData({'title': 'title', 'author': 'author'});
-  }
-
   Widget build(BuildContext context) {
-    logOut() {
-      FirebaseAuth.instance.signOut();
-    }
+    LoginViewModel viewModel = Provider.of<LoginViewModel>(context);
+    FirebaseUser currentUser = viewModel.currentUser;
 
     return ListView(
       padding: EdgeInsets.zero,
       children: <Widget>[
         UserAccountsDrawerHeader(
-          accountName: Text('UserName'),
-          accountEmail: Text('userMail@gmail.com'),
+          accountName: Text('${currentUser.displayName}'),
+          accountEmail: Text('${currentUser.email}'),
           currentAccountPicture: Image.network(
               'http://icons.iconarchive.com/icons/icons8/android/512/Users-User-icon.png'),
           decoration: BoxDecoration(color: Colors.deepOrange),
         ),
         ListTile(
           leading: Icon(Icons.history),
-          title: Text('Historique'),
-          onTap: _testDataBase,
+          title: Text('History'),
+          onTap: () {/* ... */},
         ),
         Divider(
           height: 2.0,
@@ -52,7 +46,7 @@ class Menu extends StatelessWidget {
         ListTile(
           leading: Icon(Icons.exit_to_app),
           title: Text('Logout'),
-          onTap: () => logOut(),
+          onTap: viewModel.signOut,
         )
       ],
     );
