+++ lib/src/display_ads/service/retrieve_ads.dart
@@ -2,7 +2,6 @@ import 'dart:math';
 import 'package:cest_pret_de_chez_vous/utils/list_utils.dart';
 import 'package:cest_pret_de_chez_vous/utils/random_generator.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 
 import '../utils/geo_compute.dart';
 
@@ -79,36 +78,6 @@ Future<List<Map<String, dynamic>>> getFavoriteAdsFromUser(String userId,
   List<Map<String, dynamic>> docsAsMaps = await _fetchAds(query, fromServer);
 
   return docsAsMaps;
-}
-
-Future populateTable() async {
-  var categories = ["BOOK", "KITCHEN", "ELECTRONIC"];
-  double minWest = 1.829079;
-  double maxEast = 2.516978;
-  double minSouth = 48.643868;
-  double maxNorth = 49.003630;
-  String authorId = (await FirebaseAuth.instance.currentUser()).uid;
-  final _random = new Random();
-  for (var i = 0; i < 1000; i++) {
-    String title = "Title $i";
-    String category = categories[i % 3];
-    String description =
-        "This is a random description for add $i and yes i don't know what to say, but... yeah...";
-    double latitude = minSouth + _random.nextDouble() * (maxNorth - minSouth);
-    double longitude = minWest + _random.nextDouble() * (maxEast - minWest);
-    var docRef = await Firestore().collection("ads").add({
-      "title": title,
-      "category": category,
-      "authorId": authorId,
-      "picturesUrl": [],
-      "keywords": ["keyword1"],
-      "favored": [],
-      "description": description,
-      "location": GeoPoint(latitude, longitude),
-      "createdAt": randomDate(DateTime(2015, 1), DateTime(2019, 11)),
-    });
-    docRef.updateData({"adId": docRef.documentID});
-  }
 }
 
 postNewAdToFavorites(String adId, String userId) async {
